QMimeBinaryProvider: cut out the QMap middle-man
In Qt 6, QMap is just a shared pointer to a std::map.

QMimeBinaryProvider::m_mimetypeExtra is never copied, though, so the
implicit sharing that QMap adds on top of std::map is useless.

Use the underlying std::map directly.

Yes, the std::map API is a bit raw around the edges (std::pair
value_type), but we're professionals here.

Saves more than 1.7KiB in TEXT size on optimized AMD64 GCC 11 C++20
Linux builds.

As a drive-by, cut out the insertIt temporary, too, and assign
directly to `it`.

Change-Id: If3fd98a7e7017909d00b8725f8c252b19459d5b6
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: David Faure <david.faure@kdab.com>