QLoggingRegistry: cut out the QMap middle-man
In Qt 6, QMap is just a shared pointer to a std::map.

QLoggingRegistry::qtCategoryEnvironmentOverrides is never copied,
though, so the implicit sharing that QMap adds on top of std::map is
useless.

Use the underlying std::map directly.

Yes, the std::map API is a bit raw around the edges (std::pair
value_type), but we're professionals here.

This saves more than 1.1KiB in TEXT size on optimized AMD64 GCC 11
C++20 Linux builds.

Change-Id: Id72b2432ed41a97700cc2d9ecafa902b919efe84
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>