QFactoryLoader: cut out the QMap middle-man
In Qt 6, QMap is just a shared pointer to a std::map.

QFactoryLoaderPrivate::keyMap is never copied, though, so the implicit
sharing that QMap adds on top of std::map is useless.

Use the underlying std::map directly.

Yes, the std::map API is a bit raw around the edges (std::pair
value_type), but we're professionals here.

Saves more than 1.7KiB in TEXT size on optimized AMD64 GCC 11 C++20
Linux builds.

As a drive-by, fix the double lookup in updateSinglePath().

Change-Id: I0d59f43d490fb947f653fe8dc903cf7eb9af700a
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>