Coin: Set COIN_CROSS_PLATFORM_VERSION_TEST environment variable
We have certain platforms where we build on one OS version (the
latest) and then test on an older version.
macOS is an example of such a platform.

Set the COIN_CROSS_PLATFORM_VERSION_TEST environment variable to true
for such platforms. They usually have a dependency on the main OS
version and are marked with the TestOnly feature.

This is useful information that can be used by CMake build tests in
order to skip those that do not account for the difference in OS
versions.

Change-Id: I6eeeac98abe6beb00ce431e17589f8e093ed23e4
Reviewed-by: Toni Saario <toni.saario@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>