QDataStream: inline status()
The implementation is trivial, and unchanged since the beginning of
the public project history, so I'd venture that it's safe to commit to
it not changing until Qt 7 at this point.

The reason to make it inline is that a good stream operator
implementation should be checking the stream state quite often, so we
shouldn't make that an expensive DLL entry point.

Change-Id: Iba8cbfbaf02326c86ab95be17b603cd2e785f78c
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>